bump libnDPI to 59ee1fe1156be234fed796972a29a31a0589e25a
 * set minimum nDPI version to 4.12.0 (incompatible API changes)
 * fixed `ndpi_debug_printf()` function signature
 * JSON schema (flow): added risk `56`: "Obfuscated Traffic"
 * JSON schema (flow): added "domainame"
 * fixed OpenWrt build

+++ nDPId.c
@@ -1170,7 +1170,7 @@ static void log_memory_usage(struct nDPId_reader_thread const * const reader_thr
 }
 #endif
 
-static void ndpi_debug_printf(unsigned int proto,
+static void ndpi_debug_printf(uint16_t proto,
                               struct ndpi_detection_module_struct * ndpi_struct,
                               ndpi_log_level_t log_level,
                               const char * file_name,
